requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Deprecated==1.2.13
Django==3.2.8
django-constance==2.8.0
django-cors-headers==3.10.0
django-extensions==3.1.5
django-filter==21.1
django-picklefield==3.0.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
drf-spectacular==0.20.2
drf-yasg==1.20.0
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
ipython==7.29.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
msgpack==1.0.2
multidict==5.2.0
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
pynvim==0.4.3
pyparsing==3.0.3
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
ruamel.yaml==0.17.16
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
types-pytz==2021.3.0
types-PyYAML==6.0.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
